Update SotC_PALtoNTSC.cpp
Addition of a list of folders to not copy. This includes the Bonus Material and the Translations that are not present in the NTSC version of the game.
Unfortunately, English texts replacements does not work, so I suspect the files don't have the same names in the NTSC version of the game. I need to export some of them in order to verify what's going on.